Some more work on dashing mechanic
Experimented with two options, one being where the dash ends after a set period of time, the other where the dash ended once the player's velocity zeroed out. I decided to go with the former because it was it never ended the dash early and it's easier to modify the behavior of it.